move review package under pr/ for better domain organization
internal/review → internal/pr/review
pkg/test/stub/review → pkg/test/stub/pr/review
