+++ internal/app/gui.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rin2yh/lazygh/internal/config"
 	"github.com/rin2yh/lazygh/internal/gh"
 	"github.com/rin2yh/lazygh/internal/pr/diff"
-	"github.com/rin2yh/lazygh/internal/review"
+	"github.com/rin2yh/lazygh/internal/pr/review"
 	"github.com/rin2yh/lazygh/pkg/gui/viewport"
 )
