+++ internal/app/layout/status.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/rin2yh/lazygh/internal/config"
-	"github.com/rin2yh/lazygh/internal/review"
+	"github.com/rin2yh/lazygh/internal/pr/review"
 )
 
 type Focus int
