+++ internal/pr/review/comment_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/rin2yh/lazygh/internal/config"
 	"github.com/rin2yh/lazygh/internal/gh"
 	testmock "github.com/rin2yh/lazygh/pkg/test/mock"
-	reviewstub "github.com/rin2yh/lazygh/pkg/test/stub/review"
+	reviewstub "github.com/rin2yh/lazygh/pkg/test/stub/pr/review"
 )
 
 func TestBuildDraft(t *testing.T) {
