+++ internal/pr/review/range_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rin2yh/lazygh/internal/config"
 	"github.com/rin2yh/lazygh/internal/gh"
 	testmock "github.com/rin2yh/lazygh/pkg/test/mock"
-	reviewstub "github.com/rin2yh/lazygh/pkg/test/stub/review"
+	reviewstub "github.com/rin2yh/lazygh/pkg/test/stub/pr/review"
 )
 
 func TestToggleRangeSelection_StartsAndClearsRange(t *testing.T) {
